Create README.md
By transforming the project into software or an app, it will be accessible to everyone and every person who wants to work on himself will be able to do it without any barriers. The app will be free of cost so nobody has to worry about the pricing and will be able to work with it without any restrictions.